25.11.21 - NFT Rarity code - Mob Coding/requirements.txt
python==3.9.7
Pillow==8.4.0

